refactor: move AgendaType to commonMain
