+++ androidApp/src/main/java/com/example/tasky/android/agenda/screen/AgendaDetailsScreen.kt
@@ -30,6 +30,7 @@ import androidx.navigation.NavController
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.compose.composable
 import com.example.tasky.agenda.domain.model.AgendaItem
+import com.example.tasky.agenda.domain.model.AgendaType
 import com.example.tasky.agenda.domain.model.Event
 import com.example.tasky.agenda.domain.model.Photo
 import com.example.tasky.agenda.domain.model.RemindAtType
@@ -65,12 +66,6 @@ import kotlinx.datetime.toLocalDateTime
 import kotlinx.serialization.Serializable
 import org.koin.androidx.compose.koinViewModel
 import java.time.format.DateTimeFormatter
-
-enum class AgendaType {
-    TASK,
-    EVENT,
-    REMINDER,
-}
 
 fun NavGraphBuilder.agendaDetailsScreen(navigateUp: () -> Unit) {
     composable<AgendaDetails> {
