+++ shared/src/commonMain/kotlin/com/example/tasky/agenda/domain/model/AgendaType.kt
@@ -1,0 +1,7 @@
+package com.example.tasky.agenda.domain.model
+
+enum class AgendaType {
+    TASK,
+    EVENT,
+    REMINDER,
+}
